🎨 Split client into `core`, `async` and `sync` crates.
