+++ bench/Cargo.toml
@@ -12,9 +12,11 @@ criterion = "0.3.6"
 postgres = "0.19.3"
 tokio-postgres = "0.7.6"
 postgres-types = "0.2.3"
-cornucopia = { path = "../cornucopia" }
-cornucopia_client = { path = "../cornucopia_client", features = ["sync"] }
 diesel = { version = "2.0.0-rc.0", features = ["postgres"] }
+
+cornucopia = { path = "../cornucopia" }
+cornucopia_client_sync = { path = "../clients/cornucopia_client_sync" }
+cornucopia_client_async = { path = "../clients/cornucopia_client_async" }
 
 [[bench]]
 name = "usage"
