+++ clients/cornucopia_client_async/src/generic_client.rs
@@ -169,37 +169,3 @@ impl GenericClient for Client {
         Client::query_raw(self, statement, params).await
     }
 }
-
-/// Cached statement
-pub struct Stmt {
-    query: &'static str,
-    cached: Option<Statement>,
-}
-
-impl Stmt {
-    #[must_use]
-    pub fn new(query: &'static str) -> Self {
-        Self {
-            query,
-            cached: None,
-        }
-    }
-
-    pub async fn prepare<'a, C: GenericClient>(
-        &'a mut self,
-        client: &C,
-    ) -> Result<&'a Statement, Error> {
-        if self.cached.is_none() {
-            let stmt = client.prepare(self.query).await?;
-            self.cached = Some(stmt);
-        }
-        // the statement is always prepared at this point
-        Ok(unsafe { self.cached.as_ref().unwrap_unchecked() })
-    }
-}
-
-/// This trait allows you to bind parameters to a query using a single
-/// struct, rather than passing each bind parameter as a function parameter.
-pub trait Params<'a, P, O, C> {
-    fn params(&'a mut self, client: &'a C, params: &'a P) -> O;
-}
