+++ clients/cornucopia_client_core/src/array_iterator.rs
@@ -1,9 +1,10 @@
 use fallible_iterator::FallibleIterator;
 use postgres_protocol::types::{array_from_sql, ArrayValues};
 use postgres_types::{FromSql, Kind, Type};
-use std::{fmt::Debug, marker::PhantomData};
+use std::fmt::Debug;
+use std::marker::PhantomData;
 
-use crate::private::escape_domain;
+use crate::utils::escape_domain;
 
 /// Iterator over the items in a PostgreSQL array. You only need this if you are
 /// working with custom zero-cost type mapping of rows containing PostgreSQL arrays.
