+++ clients/cornucopia_client_core/src/lib.rs
@@ -1,0 +1,7 @@
+mod array_iterator;
+mod domain;
+mod utils;
+
+pub use array_iterator::ArrayIterator;
+pub use domain::{Domain, DomainArray};
+pub use utils::slice_iter;
