+++ clients/cornucopia_client_sync/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "cornucopia_client_sync"
+version = "0.2.2"
+edition = "2021"
+license = "MIT/Apache-2.0"
+description = "Generic client trait for Cornucopia users"
+homepage = "https://github.com/cornucopia-rs/cornucopia"
+repository = "https://github.com/cornucopia-rs/cornucopia"
+readme = "README.md"
+categories = ["database"]
+keywords = ["postgresql", "query", "generator", "sql", "tokio-postgres"]
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+postgres = { version = "0.19.3" }
+cornucopia_client_core = { path = "../cornucopia_client_core" }
