+++ codegen_test/Cargo.toml
@@ -18,7 +18,6 @@ tokio-postgres = { version = "0.7.6", features = [
     "with-uuid-1",
     "with-eui48-1",
 ] }
-cornucopia_client = { path = "../cornucopia_client", features = ["sync"] }
 postgres-types = { version = "0.2.3", features = ["derive"] }
 serde = { version = "1.0.138", features = ["derive"] }
 serde_json = { version = "1.0.82", features = ["raw_value"] }
@@ -26,3 +25,6 @@ time = { version = "0.3.11", features = ["parsing", "serde"] }
 uuid = { version = "1.1.2", features = ["serde"] }
 eui48 = { version = "1.1.0", features = ["serde"] }
 futures = "0.3.21"
+
+cornucopia_client_sync = { path = "../clients/cornucopia_client_sync" }
+cornucopia_client_async = { path = "../clients/cornucopia_client_async" }
