+++ examples/auto_build/Cargo.toml
@@ -8,10 +8,11 @@ edition = "2021"
 [dependencies]
 tokio = { version = "1.18.2", features = ["full"] }
 tokio-postgres = "0.7.6"
-cornucopia_client = { path = "../../cornucopia_client" }
 deadpool-postgres = "0.10.2"
 futures = "0.3.21"
 postgres-types = "0.2.3"
+
+cornucopia_client_async = { path = "../../clients/cornucopia_client_async" }
 
 [build-dependencies]
 cornucopia = { path = "../../cornucopia" }
